the square root of a number

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -1,11 +1,11 @@
 #include "main.h"
-int squareroot(int n, int i);
+int squareroot(long int n, long int i);
 /**
  * _sqrt_recursion - no loop, 1. 2nd function - checks for perfect square
  * @n: input
  * Return: Always 0 (Success)
  */
-int _sqrt_recursion(int n)
+int _sqrt_recursion(long int n)
 {
 	if (n < 0)
 		return (-1);
@@ -18,7 +18,7 @@ int _sqrt_recursion(int n)
  * @i: counter
  * Return: if square root
  */
-int squareroot(int n, int i)
+int squareroot(long int n, long int i)
 {
 	if (i < 1)
 		return (-1);
